adding the AdminController
moving the functions related to the admin from the userController to the AdminController
